requirements.txt
beautifulsoup4
scraperapi-sdk
twilio

